backend/requirements.txt
# DIRK Brain Portal - Backend Dependencies

# Core Framework
fastapi==0.116.1
uvicorn[standard]==0.35.0
python-multipart==0.0.9

# WebSocket & Real-time
websockets==15.0.1
aiohttp==3.12.15

# gRPC
grpcio==1.74.0
grpcio-tools==1.74.0
protobuf==6.31.1

# Database
sqlalchemy==2.0.23
alembic==1.13.1
asyncpg==0.29.0
psycopg2-binary==2.9.9

# Security & Validation
pydantic==2.11.7
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# AI/ML Libraries (optional)
openai==1.54.3
anthropic==0.39.0
google-generativeai==0.8.3

# Utilities
httpx==0.27.2
python-dateutil==2.9.0
pyyaml==6.0.1
click==8.2.1

# Development & Testing
pytest==8.3.3
pytest-asyncio==0.24.0
black==24.10.0
flake8==7.1.1
mypy==1.13.0

# Monitoring
prometheus-client==0.21.0
opentelemetry-api==1.28.2
opentelemetry-sdk==1.28.2

# Task Queue (optional)
celery==5.4.0
redis==5.2.0